Move demos into /demo folder
