+++ usellm.org/content/docs/api-reference/chat.mdx
@@ -33,7 +33,7 @@ The `chat` function returns a `Promise` that resolves to an `LLMChatResult` obje
 
 ### Example
 
-Here is an example usage of the `chat` function within a React component ([live demo](/demo1)):
+Here is an example usage of the `chat` function within a React component ([live demo](/demo/llm-chat)):
 
 ```jsx
 "use client";
