+++ usellm.org/content/docs/api-reference/record.mdx
@@ -35,7 +35,7 @@ The `stopRecording` function returns a `Promise` that resolves to an object cont
 
 ### Example
 
-Here is an example usage of the `record` and `stopRecording`([live demo](/demos/audio)):
+Here is an example usage of the `record` and `stopRecording`([live demo](/demo/speech-to-text)):
 
 ```jsx
 'use client';
