+++ usellm.org/content/docs/api-reference/register-template.mdx
@@ -39,7 +39,7 @@ llmService.registerTemplate({
 
 ### Example
 
-Here's an example of how you can use this method in a Next.js [app router](https://nextjs.org/docs/app) project ([live demo](https://usellm.org/demo2)):
+Here's an example of how you can use this method in a Next.js [app router](https://nextjs.org/docs/app) project ([live demo](https://usellm.org/demo/tutorial-generator)):
 
 ```javascript
 /* app/api/llm.tsx */
