+++ usellm.org/content/docs/api-reference/usellm-hook.mdx
@@ -38,7 +38,7 @@ An object with the following functions:
 
 ### Example
 
-Here's a complete working example that you can use as a starting point ([live demo](/demo1)):
+Here's a complete working example that you can use as a starting point ([live demo](/demo/llm-chat)):
 
 ```javascript
 "use client";
