+++ usellm.org/content/docs/contribute/contribution-guide.mdx
@@ -81,7 +81,7 @@ tested on the `usellm.org` application.
 Here are some good places to start:
 
 - Improve the documentation by making changes in the folder `usellm.org/content/docs` (edit the `.mdx` files)
-- Add or improve/modify demo applications within the `usellm.org/app/demos` folder
+- Add or improve/modify demo applications within the `usellm.org/app/demo` folder
 - Fix bugs or add new features to the `usellm` package within the `packages/usellm` folder
 - Browse through open issues on the [useLLM repository](https://github.com/usellm/usellm/issues) and see if you can help out!
