+++ usellm.org/content/docs/index.mdx
@@ -3,7 +3,7 @@ title: Quickstart
 description: Add large language models to your React app with just a few lines of code.
 ---
 
-`useLLM` is a React hook for integrating large language models like OpenAI's ChatGPT with just a few lines of code. [Try Demo](https://usellm.org/demo)
+`useLLM` is a React hook for integrating large language models like OpenAI's ChatGPT with just a few lines of code. [Try Demo](https://usellm.org/demo/chat-thread)
 
 ## Installation
 
@@ -66,7 +66,7 @@ The `messages` option passed to `llm.chat` must be an array of OpenAI messages, 
 
 #### Example - `llm.chat`
 
-Here's a complete working example that you can use as a starting point ([live demo](/demo1)):
+Here's a complete working example that you can use as a starting point ([live demo](/demo/llm-chat)):
 
 ```javascript
 "use client";
@@ -328,7 +328,7 @@ Check out this course on prompt engineering to craft effective prompts: https://
 
 #### Example - Template
 
-Here's a complete example of using templates in a NextJS [app router](https://nextjs.org/docs/app) project ([live demo](https://usellm.org/demo2)):
+Here's a complete example of using templates in a NextJS [app router](https://nextjs.org/docs/app) project ([live demo](https://usellm.org/demo/tutorial-generator)):
 
 ```javascript
 /* app/api/llm.tsx */
